server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
bert-score==0.3.13
    # via -r requirements.in
blis==1.3.3
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer-slim
    #   uvicorn
cloudpathlib==0.23.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
cymem==2.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
datasets==2.14.4
    # via
    #   -r requirements.in
    #   evaluate
dill==0.3.7
    # via
    #   datasets
    #   evaluate
    #   multiprocess
distro==1.9.0
    # via openai
evaluate==0.4.6
    # via -r requirements.in
fastapi==0.127.0
    # via -r requirements.in
filelock==3.20.3
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.61.1
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2026.1.0
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   torch
fuzzywuzzy==0.18.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via openai
huggingface-hub==0.36.0
    # via
    #   datasets
    #   evaluate
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   spacy
    #   torch
jiter==0.12.0
    # via openai
joblib==1.5.3
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
levenshtein==0.27.3
    # via python-levenshtein
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via bert-score
mpmath==1.3.0
    # via sympy
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via
    #   datasets
    #   evaluate
murmurhash==1.0.15
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==3.6.1
    # via torch
numpy==2.4.1
    # via
    #   bert-score
    #   blis
    #   contourpy
    #   datasets
    #   evaluate
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
openai==2.15.0
    # via -r requirements.in
packaging==25.0
    # via
    #   bert-score
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   matplotlib
    #   spacy
    #   thinc
    #   transformers
    #   weasel
pandas==2.3.3
    # via
    #   -r requirements.in
    #   bert-score
    #   datasets
    #   evaluate
pillow==12.1.0
    # via matplotlib
preshed==3.0.12
    # via
    #   spacy
    #   thinc
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyarrow==23.0.0
    # via datasets
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   confection
    #   fastapi
    #   openai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.41.5
    # via pydantic
pymupdf==1.26.7
    # via -r requirements.in
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   uvicorn
python-levenshtein==0.27.3
    # via -r requirements.in
python-multipart==0.0.21
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
    #   uvicorn
rapidfuzz==3.14.3
    # via levenshtein
regex==2026.1.15
    # via
    #   tiktoken
    #   transformers
requests==2.32.5
    # via
    #   bert-score
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   spacy
    #   tiktoken
    #   transformers
    #   weasel
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smart-open==7.5.0
    # via weasel
sniffio==1.3.1
    # via openai
spacy==3.8.11
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.2
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
starlette==0.50.0
    # via fastapi
sympy==1.14.0
    # via torch
thinc==8.3.10
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.12.0
    # via -r requirements.in
tokenizers==0.22.2
    # via transformers
torch==2.9.1
    # via
    #   bert-score
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   bert-score
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   spacy
    #   transformers
transformers==4.57.6
    # via
    #   -r requirements.in
    #   bert-score
    #   sentence-transformers
triton==3.5.1
    # via torch
typer-slim==0.21.1
    # via
    #   spacy
    #   weasel
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   starlette
    #   torch
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.40.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
watchfiles==1.1.1
    # via uvicorn
weasel==0.4.3
    # via spacy
websockets==15.0.1
    # via uvicorn
wrapt==2.0.1
    # via smart-open
xxhash==3.6.0
    # via
    #   datasets
    #   evaluate
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
